unified_studio_app/requirements.txt
streamlit==1.39.0
numpy==1.26.4
pillow==10.4.0
anthropic>=0.25.0
openai>=1.40.0
